get analysis code working
